Update InstrumentoCordas
Now instead of showing the chords directly, an array is created in which 0's are not-pressed notes and 1's are pressed notes. After the arrayis created, the chord is show the old way, but need the array to work. Now I should focus on displaying the note name and spaces numeration of the shown chord. then I should proceed to apply the flipping algorithm